part_1

+++ internal/clients/accural.go
@@ -22,9 +22,9 @@ var (
 )
 
 type OrderResponse struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int64  `json:"accrual,omitempty"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual,omitempty"`
 }
 
 type LoyaltyClient struct {
